requirements.txt
# Global requirements for Databricks IQ project
# This file combines dependencies for both UI and MCP server components

# Core Python packages
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.0.0
requests>=2.28.0
python-multipart>=0.0.6

# Streamlit UI dependencies
streamlit>=1.28.0
altair>=5.0.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Databricks SDK and connectors
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0

# MCP Server dependencies
fastmcp>=0.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncio-extras>=1.3.0

# MCP protocol
mcp[cli]>=0.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0